Refine icon props and location parsing logic
Standardized icon props in useSectionIcons and adjusted default icon size. Updated useWishlist to use parseFloat instead of parseInt for latitude and longitude parsing, ensuring more accurate geolocation. Minor className and padding adjustments in FactsSection and FAQ components for improved UI consistency.

+++ hooks/wishlist/use-wishlist.ts
@@ -36,11 +36,11 @@ const useWishlist = () => {
       radius: "20",
       lat: latLng
         ? latLng.lat
-        : parseInt(LocationHelper.getLocation().latitude ?? "40.7127753") ||
+        : parseFloat(LocationHelper.getLocation().latitude ?? "40.7127753") ||
           40.7127753,
       lon: latLng
         ? latLng.lng
-        : parseInt(LocationHelper.getLocation().longitude ?? "-74.0059728") ||
+        : parseFloat(LocationHelper.getLocation().longitude ?? "-74.0059728") ||
           -74.0059728,
       careType: "Skilled Nursing",
       page: 1,
